+++ src/Console/Command/DockerComposeCommand.php
@@ -23,7 +23,7 @@ class DockerComposeCommand extends Command
 
     /** @var array */
     private $allowedCommands = array(
-        'build', 'kill', 'logs', 'port', 'ps', 'pull', 'rm', 'scale', 'start', 'stop', 'up'
+        'build', 'kill', 'logs', 'ps', 'pull', 'rm', 'scale', 'start', 'stop', 'up'
     );
 
     /**